requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.3
openpyxl==3.1.2
dashscope==1.17.0
python-dotenv==1.0.0

